Django/osc_site/requirements.txt
Django
django_cron
requests
polyline
elasticsearch
slacker
psycopg2
setuptools
matplotlib
pyproj
pandas
pyshp
feedparser
geopy
pytz
djangorestframework
